refactor FolderListTile & ReceiptListTile to separate files
